Move AppInitializers to common code
- Add the initialize to the iOS component and implement it in the AppDelegate on iOS code.
