+++ core/base/src/commonMain/kotlin/com/thomaskioko/tvmaniac/core/base/AppInitializers.kt
@@ -1,6 +1,5 @@
-package com.thomaskioko.tvmaniac.initializers
+package com.thomaskioko.tvmaniac.core.base
 
-import com.thomaskioko.tvmaniac.core.base.AppInitializer
 import software.amazon.lastmile.kotlin.inject.anvil.AppScope
 import me.tatarka.inject.annotations.Inject
 import software.amazon.lastmile.kotlin.inject.anvil.SingleIn
@@ -10,7 +9,7 @@ import software.amazon.lastmile.kotlin.inject.anvil.SingleIn
 class AppInitializers(
   private val initializers: Set<AppInitializer>,
 ) {
-  fun init() {
+  fun initialize() {
     for (initializer in initializers) {
       initializer.init()
     }
